Add a script to check invalid screenshot(s).
