+++ .github/workflows/quality.yml
@@ -21,6 +21,18 @@ jobs:
       - name: Run code quality check suite
         run: ./tools/check/check_code_quality.sh
 
+  checkScreesnhot:
+    name: Search for invalid screenshot files
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python 3.9
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.9
+      - name: Search for invalid screenshot files
+        run: ./tools/test/checkInvalidScreenshots.py
+
   check:
     name: Project Check Suite
     runs-on: ubuntu-latest
